requirements.txt
# Tested on Python 3.13.7
numpy
scipy
matplotlib
imageio
PyWavelets
numba
streamlit
Pillow